Move RainMachine service registration to async_setup
Register the RainMachine services in async_setup instead of
async_setup_entry so they are available for automation validation even
when no config entry is loaded. Service handlers now resolve the target
controller via call.hass and raise ServiceValidationError for an invalid
device ID, a non-RainMachine device, or a controller that is not loaded.

The service code moves to a dedicated services.py, and
async_update_programs_and_zones moves to coordinator.py to avoid a
circular import. This removes the need to deregister services on unload
and the home-assistant-service-registered-in-setup-entry pylint disable.

Add tests covering service registration, each service call, controller
errors, and the ServiceValidationError paths.
